backend/pyproject.toml
[tool.poetry]
name = "admin-backend"
version = "0.1.0"
description = "通用管理后台系统"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.23"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
email-validator = "^2.1.0.post1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi-line-output = 3 